+++ src/StepSeqPlus.cpp
@@ -222,22 +222,71 @@ struct StepSeqPlus : Module {
 
 		configInput(LENGTH_INPUT, "Length");
 
-		configParam(STEP_PARAM+0, 0, 1.f, 0.5f, "Step 1");
-		configParam(STEP_PARAM+1, 0, 1.f, 0.5f, "Step 2");
-		configParam(STEP_PARAM+2, 0, 1.f, 0.5f, "Step 3");
-		configParam(STEP_PARAM+3, 0, 1.f, 0.5f, "Step 4");
-		configParam(STEP_PARAM+4, 0, 1.f, 0.5f, "Step 5");
-		configParam(STEP_PARAM+5, 0, 1.f, 0.5f, "Step 6");
-		configParam(STEP_PARAM+6, 0, 1.f, 0.5f, "Step 7");
-		configParam(STEP_PARAM+7, 0, 1.f, 0.5f, "Step 8");
-		configParam(STEP_PARAM+8, 0, 1.f, 0.5f, "Step 9");
-		configParam(STEP_PARAM+9, 0, 1.f, 0.5f, "Step 10");
-		configParam(STEP_PARAM+10, 0, 1.f, 0.5f, "Step 11");
-		configParam(STEP_PARAM+11, 0, 1.f, 0.5f, "Step 12");
-		configParam(STEP_PARAM+12, 0, 1.f, 0.5f, "Step 13");
-		configParam(STEP_PARAM+13, 0, 1.f, 0.5f, "Step 14");
-		configParam(STEP_PARAM+14, 0, 1.f, 0.5f, "Step 15");
-		configParam(STEP_PARAM+15, 0, 1.f, 0.5f, "Step 16");
+		struct RangeQuantity : ParamQuantity {
+			float getDisplayValue() override {
+				StepSeqPlus* module = reinterpret_cast<StepSeqPlus*>(this->module);
+
+				switch (module->range) {
+					case 0:	// 0/1v
+						displayMultiplier = 1.f;
+						displayOffset = 0.f;
+					break;
+					case 1:	// 0/2v
+						displayMultiplier = 2.f;
+						displayOffset = 0.f;
+					break;
+					case 2:	// 0/3v
+						displayMultiplier = 3.f;
+						displayOffset = 0.f;
+					break;
+					case 3:	// 0/5v
+						displayMultiplier = 5.f;
+						displayOffset = 0.f;
+					break;
+					case 4:	// 0/10v
+						displayMultiplier = 10.f;
+						displayOffset = 0.f;
+					break;
+					case 5:	// -1/+1v
+						displayMultiplier = 2.f;
+						displayOffset = -1.f;
+					break;
+					case 6:	// -2/+2v
+						displayMultiplier = 4.f;
+						displayOffset = -2.f;
+					break;
+					case 7:	// -3/+3v
+						displayMultiplier = 6.f;
+						displayOffset = -3.f;
+					break;
+					case 8:	// -5/+5v
+						displayMultiplier = 10.f;
+						displayOffset = -5.f;
+					break;
+					case 9:	// -10/+10v
+						displayMultiplier = 20.f;
+						displayOffset = -10.f;
+					break;
+				}
+				return ParamQuantity::getDisplayValue();
+			}
+		};
+		configParam<RangeQuantity>(STEP_PARAM+0, 0, 1.f, 0.5f, "Step 1");
+		configParam<RangeQuantity>(STEP_PARAM+1, 0, 1.f, 0.5f, "Step 2");
+		configParam<RangeQuantity>(STEP_PARAM+2, 0, 1.f, 0.5f, "Step 3");
+		configParam<RangeQuantity>(STEP_PARAM+3, 0, 1.f, 0.5f, "Step 4");
+		configParam<RangeQuantity>(STEP_PARAM+4, 0, 1.f, 0.5f, "Step 5");
+		configParam<RangeQuantity>(STEP_PARAM+5, 0, 1.f, 0.5f, "Step 6");
+		configParam<RangeQuantity>(STEP_PARAM+6, 0, 1.f, 0.5f, "Step 7");
+		configParam<RangeQuantity>(STEP_PARAM+7, 0, 1.f, 0.5f, "Step 8");
+		configParam<RangeQuantity>(STEP_PARAM+8, 0, 1.f, 0.5f, "Step 9");
+		configParam<RangeQuantity>(STEP_PARAM+9, 0, 1.f, 0.5f, "Step 10");
+		configParam<RangeQuantity>(STEP_PARAM+10, 0, 1.f, 0.5f, "Step 11");
+		configParam<RangeQuantity>(STEP_PARAM+11, 0, 1.f, 0.5f, "Step 12");
+		configParam<RangeQuantity>(STEP_PARAM+12, 0, 1.f, 0.5f, "Step 13");
+		configParam<RangeQuantity>(STEP_PARAM+13, 0, 1.f, 0.5f, "Step 14");
+		configParam<RangeQuantity>(STEP_PARAM+14, 0, 1.f, 0.5f, "Step 15");
+		configParam<RangeQuantity>(STEP_PARAM+15, 0, 1.f, 0.5f, "Step 16");
 
 		configParam(PROG_PARAM, 0.f, 31.f, 0.f, "Prog");
 		configInput(PROG_INPUT, "Prog");
@@ -1250,7 +1299,8 @@ struct StepSeqPlusWidget : ModuleWidget {
 	void appendContextMenu(Menu* menu) override {
 		StepSeqPlus* module = dynamic_cast<StepSeqPlus*>(this->module);
 
-		
+		menu->addChild(new MenuSeparator());
+
 		struct RangeItem : MenuItem {
 			StepSeqPlus* module;
 			int range;
@@ -1259,7 +1309,6 @@ struct StepSeqPlusWidget : ModuleWidget {
 			}
 		};
 
-		menu->addChild(new MenuSeparator());
 		std::string rangeNames[10] = {"0/1v", "0/2v", "0/3v", "0/5v", "0/10v", "-1/+1v", "-2/+2v", "-3/+3v", "-5/+5v", "-10/+10v"};
 		menu->addChild(createSubmenuItem("Knobs Range", rangeNames[module->range], [=](Menu * menu) {
 			for (int i = 0; i < 10; i++) {
@@ -1271,6 +1320,8 @@ struct StepSeqPlusWidget : ModuleWidget {
 			}
 		}));
 		
+		menu->addChild(new MenuSeparator());
+		
 		struct RunTypeItem : MenuItem {
 			StepSeqPlus* module;
 			int runType;
@@ -1279,16 +1330,16 @@ struct StepSeqPlusWidget : ModuleWidget {
 			}
 		};
 
-		menu->addChild(new MenuSeparator());
-		menu->addChild(createMenuLabel("Run Input"));
 		std::string RunTypeNames[2] = {"Gate", "Trig"};
-		for (int i = 0; i < 2; i++) {
-			RunTypeItem* runTypeItem = createMenuItem<RunTypeItem>(RunTypeNames[i]);
-			runTypeItem->rightText = CHECKMARK(module->runType == i);
-			runTypeItem->module = module;
-			runTypeItem->runType = i;
-			menu->addChild(runTypeItem);
-		}
+		menu->addChild(createSubmenuItem("Run Input", (RunTypeNames[module->runType]), [=](Menu * menu) {
+			for (int i = 0; i < 2; i++) {
+				RunTypeItem* runTypeItem = createMenuItem<RunTypeItem>(RunTypeNames[i]);
+				runTypeItem->rightText = CHECKMARK(module->runType == i);
+				runTypeItem->module = module;
+				runTypeItem->runType = i;
+				menu->addChild(runTypeItem);
+			}
+		}));
 
 		struct RevTypeItem : MenuItem {
 			StepSeqPlus* module;
@@ -1297,17 +1348,16 @@ struct StepSeqPlusWidget : ModuleWidget {
 				module->revType = revType;
 			}
 		};
-
-		menu->addChild(new MenuSeparator());
-		menu->addChild(createMenuLabel("Reverse Input Voltage"));
 		std::string RevTypeNames[2] = {"Positive", "Negative"};
-		for (int i = 0; i < 2; i++) {
-			RevTypeItem* revTypeItem = createMenuItem<RevTypeItem>(RevTypeNames[i]);
-			revTypeItem->rightText = CHECKMARK(module->revType == i);
-			revTypeItem->module = module;
-			revTypeItem->revType = i;
-			menu->addChild(revTypeItem);
-		}
+		menu->addChild(createSubmenuItem("Reverse Input Voltage", (RevTypeNames[module->revType]), [=](Menu * menu) {
+			for (int i = 0; i < 2; i++) {
+				RevTypeItem* revTypeItem = createMenuItem<RevTypeItem>(RevTypeNames[i]);
+				revTypeItem->rightText = CHECKMARK(module->revType == i);
+				revTypeItem->module = module;
+				revTypeItem->revType = i;
+				menu->addChild(revTypeItem);
+			}
+		}));
 
 		menu->addChild(new MenuSeparator());
 		menu->addChild(createBoolPtrMenuItem("Reset on Run", "", &module->rstOnRun));
@@ -1328,10 +1378,6 @@ struct StepSeqPlusWidget : ModuleWidget {
 		menu->addChild(createMenuItem("Save PROG preset", "", [=]() {module->menuSavePreset();}));
 
 		menu->addChild(new MenuSeparator());
-		menu->addChild(createMenuLabel("Store Programs"));
-		menu->addChild(createMenuLabel("with double-click"));
-
-		menu->addChild(new MenuSeparator());
 		menu->addChild(createSubmenuItem("Erase ALL progs", "", [=](Menu * menu) {
 			menu->addChild(createSubmenuItem("Are you Sure?", "", [=](Menu * menu) {
 				menu->addChild(createMenuItem("ERASE!", "", [=]() {module->eraseProgs();}));
@@ -1340,6 +1386,11 @@ struct StepSeqPlusWidget : ModuleWidget {
 
 		menu->addChild(new MenuSeparator());
 		menu->addChild(createBoolPtrMenuItem("Initialize on Start", "", &module->initStart));
+
+		menu->addChild(new MenuSeparator());
+		menu->addChild(createSubmenuItem("Hints", "", [=](Menu * menu) {
+			menu->addChild(createMenuLabel("Store Programs with double-click"));
+		}));
 	}
 
 };